Fix off-by-one stack pointer
This is the root cause of 2 weeks of bug hunting. Instead of
following the official ABI of "sp points at the last used value" it was
instead previously written that it pointed at the first free. This was
fine until we added C then it became an issue because of mismatched ABI
calling convention. The bug led to stack corruption during context
switches.